Moved figure/layer setup and callback management to pca_disc backend.
